refactor: move IDBStore enum to idb manager and update imports across the application
