+++ apps/web/src/routes/my-kiwis/-apis/get-kiwis.ts
@@ -1,5 +1,4 @@
-import { IDBStore } from "#/constants/idb";
-import idb from "#/managers/idb";
+import idb, { IDBStore } from "#/managers/idb";
 import { KiwiIDBData } from "#/types/idb";
 import { Kiwi } from "#/types/kiwi";
 import { kiwIDBDataToKiwi } from "#/utils/idb";
